refactor: move web/ and channel_adapter.rs into channels/ module
- Move src/web/ -> src/channels/web/
- Move src/channel_adapter.rs -> src/channels/adapter.rs
- Update all crate::web -> crate::channels::web references
- Update all crate::channel_adapter -> crate::channels::adapter references
