+++ src/channels/adapter.rs
@@ -113,7 +113,7 @@ mod tests {
     use std::sync::Arc;
 
     use super::*;
-    use crate::web::WebAdapter;
+    use crate::channels::web::WebAdapter;
 
     #[test]
     fn registry_registers_and_resolves() {
