+++ src/lib.rs
@@ -5,27 +5,20 @@
 
 pub mod agent_loop;
 pub mod assets;
-pub mod channel_adapter;
 pub mod channels;
 pub mod codex_auth;
 pub mod config;
 pub mod error;
-pub mod gateway;
 pub mod llm;
 pub mod llm_profile;
-pub mod logging;
 pub mod mcp;
-pub mod media;
 pub mod runtime;
 pub mod setup;
 pub mod skills;
 pub mod slash_commands;
 pub mod soul_agents;
-pub mod status;
 pub mod storage;
-pub mod text;
 pub mod tools;
-pub mod web;
 
 #[cfg(test)]
 mod test_env;
